pyproject.toml
[project]
name            = "hnbot"
version         = "0.0.0"
description     = ""
readme          = "README.md"
requires-python = ">=3.12"
authors         = [{ name = "narumi", email = "toucans-cutouts0f@icloud.com" }]
dependencies    = [
  "aiocache[redis]>=0.12.3",
  "aiogram>=3.26.0",
  "feedparser>=6.0.12",
  "httpx>=0.28.1",
  "logfire>=4.31.0",
  "loguru>=0.7.3",
  "markdownify>=1.2.2",
  "openai>=2.30.0",
  "pydantic>=2.12.5",
  "python-dotenv>=1.2.2",
  "rich>=14.3.2",
  "telegraph>=2.2.0",
  "typer>=0.21.1",
]

[project.scripts]
hnbot = "hnbot.cli:app"

[dependency-groups]
dev = [
  "pytest>=9.0.2",
  "pytest-cov>=7.0.0",
  "ruff>=0.15.0",
  "ty>=0.0.15",
]

[build-system]
requires      = ["uv_build>=0.8.17,<0.9.0"]
build-backend = "uv_build"

[tool.bumpversion]
current_version  = "0.0.0"
tag              = true
commit           = true
pre_commit_hooks = ["uv lock", "git add uv.lock"]

[tool.pytest]
filterwarnings = ["ignore::DeprecationWarning"]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
  "ANN",     # flake8-annotations
  "ASYNC",   # flake8-async
  "B",       # flake8-bugbear
  "BLE",     # flake8-blind-except
  "C",       # flake8-comprehensions
  "DTZ",     # flake8-datetimez
  "E",       # pycodestyle errors
  "F",       # pyflakes
  "FURB",    # refurb
  "I",       # isort
  "G",       # flake8-logging-format
  "N",       # pep8-naming
  "PERF",    # perflint
  "PIE",     # flake8-pie
  "PTH",     # flake8-use-pathlib
  "RET",     # flake8-return
  "RSE",     # flake8-raise
  "RUF",     # ruff-specific rules
  "SIM",     # flake8-simplify
  "TID",     # flake8-tidy-imports
  "TRY004",  # type-check-without-type-error
  "TRY203",  # useless-try-except
  "TRY300",  # try-consider-else
  "UP",      # pyupgrade
  "W",       # pycodestyle warnings
]

[tool.ruff.lint.per-file-ignores]
"tests/**.py" = ["ANN", "TRY", "PERF"]

[tool.ruff.lint.isort]
force-single-line = true

[tool.tombi.format.rules]
key-value-equals-sign-alignment = true
trailing-comment-alignment      = true

[tool.ty.rules]
unresolved-import = "ignore"
